# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phonebook API",
        "description": "API documentation for the Phonebook",
        "version": "1.0.0"
    },
    "paths": {
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get list of contacts",
                "description": "Returns list of contacts",
                "operationId": "2dd193c75a66de01c11e857e6cb14cbd",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a new contact",
                "description": "Creates a new contact",
                "operationId": "c2b1ce43f7cbbb90d954de7f941e5e3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created"
                    }
                }
            }
        },
        "/api/contacts/{contact_id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get a specific contact",
                "description": "Returns a single contact",
                "operationId": "cd6b7e789590dc81f346c6a0d76bb5dd",
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Update a specific contact",
                "description": "Updates a single contact",
                "operationId": "bc1a6058c70005f2b66491e592090537",
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Delete a specific contact",
                "description": "Deletes a single contact",
                "operationId": "4380782bd8f9e3fe1fba27aba33b87b9",
                "parameters": [
                    {
                        "name": "contact_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "title": "Contact",
                "required": [
                    "name",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "integer",
                        "example": "1234567890"
                    },
                    "marked": {
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Contacts",
            "description": "API Endpoints for Contacts"
        }
    ]
}